chore: move '_check_batch_status' to only calling module
Likewise the 'FauxCall' helper class it uses.
